General Tasks Detachment
Now General Tasks are part of the core library
