--- include/Elpida/CMakeLists.txt
@@ -1,7 +1,0 @@
-configure_file(Config.hpp.in Config.hpp)
-
-add_library(Elpida CpuFeature.cpp)
-
-target_include_directories(Elpida PUBLIC
-        "${PROJECT_BINARY_DIR}"
-        )
