--- include/TaskBatches/General/meson.build
@@ -1,50 +1,0 @@
-#-------------------------------------------------------------------------------
-# elpida - CPU benchmark tool
-# 
-# Copyright (C) 2018  Ioannis Panagiotopoulos
-# 
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-# 
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-# 
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>
-#-------------------------------------------------------------------------------
-
-generalTasks_sources = ['ReadFile.cpp', 
-						'WriteFile.cpp',
-						'Plugin.cpp',
-						'AllocateMemory.cpp',
-						'Memory.cpp',
-						'AlignedMemory.cpp',
-						'UnalignedMemory.cpp',
-						'MemoryChunk.cpp',
-						'NumaMemory.cpp',
-						'NumaAllocatePerThread.cpp']
-
-generalTasks = library('GeneralTasks', 
-							generalTasks_sources,
-							link_with: elpida,
-							install: true,
-							dependencies: [numa_dep],
-							cpp_args: defaultCppArgs,
-							install_dir: pluginInstallDir,
-							include_directories : rootIncude)
-							
-install_headers(['ReadFile.hpp',
-				'WriteFile.hpp',
-				'AllocateMemory.hpp',
-				'Memory.hpp',
-				'AlignedMemory.hpp',
-				'UnalignedMemory.hpp',
-				'MemoryChunk.hpp',
-				'NumaMemory.cpp',
-				'NumaAllocatePerThread.cpp'],
-				subdir: 'Elpida/TaskBatches/General')
-
