+++ src/TaskBatches/Image/ImageTaskBatch.cpp
@@ -31,8 +31,8 @@
 #include "Elpida/Utilities/CommandParser.hpp"
 #include "Elpida/Utilities/Image.hpp"
 #include "Elpida/Utilities/MemoryFile.hpp"
-#include "TaskBatches/General/ReadFile.hpp"
-#include "TaskBatches/General/WriteFile.hpp"
+#include "Elpida/CommonTasks/ReadFile.hpp"
+#include "Elpida/CommonTasks/WriteFile.hpp"
 #include "TaskBatches/Image/ConvertToFloat.hpp"
 #include "TaskBatches/Image/ConvertToUInt8.hpp"
 #include "TaskBatches/Image/FloydSteinberg.hpp"
