--- src/TaskBatches/plugins
@@ -1,3 +1,0 @@
-GeneralTasks
-MemoryTasks
-ImageTasks
